+++ src/app/payment-hub/request-to-pay/helper/ams-short-codes.ts
@@ -1,0 +1,9 @@
+import { environment } from "environments/environment";
+
+export const amsShortCodes = (type: any = undefined) => {
+    let amsCodes = environment.amsShortCodes;
+    if(type){
+        return amsCodes.filter(code => type === code.type);
+    }
+    return amsCodes;
+};