+++ src/app/payment-hub/request-to-pay/incoming-request-export/incoming-request-export.component.ts
@@ -1,23 +1,24 @@
 import { Component, OnInit } from "@angular/core";
 import { RequestToPayService } from "../service/request-to-pay.service";
+import { amsShortCodes } from "../helper/ams-short-codes";
+
 @Component({
   selector: "mifosx-incoming-request-export",
   templateUrl: "./incoming-request-export.component.html",
   styleUrls: ["./incoming-request-export.component.scss"],
 })
 export class IncomingRequestExportComponent implements OnInit {
+  amsCodes = amsShortCodes('TILL');
   csvExport: [];
   csvName: string;
   constructor(private requestToPayService: RequestToPayService) {}
 
   ngOnInit(): void {}
-  onSubmit() {
-    this.exportCSV(this.csvExport, this.csvName);
-  }
+
   arrayConvert(event: any) {
     event.target.value.split(",");
   }
-  exportCSV(filterBy: any, filterName: string) {
-    this.requestToPayService.exportCSV(filterBy, filterName);
+  exportCSV(filterBy: any) {
+    this.requestToPayService.exportCSV(filterBy);
   }
 }