+++ src/app/system/audit-trails/audit-trail.datasource.ts
@@ -38,7 +38,7 @@ export class AuditTrailsDataSource implements DataSource<any> {
     this.systemService.getAuditTrails(filterBy, orderBy, sortOrder, pageIndex * limit, limit)
       .subscribe((auditTrails: any) => {
         this.recordsSubject.next(auditTrails.totalFilteredRecords);
-        this.auditTrailsSubject.next(auditTrails.pageItems);
+        this.auditTrailsSubject.next(auditTrails.content);
       });
   }