+++ RELEASE_NOTES.md
@@ -1,5 +1,33 @@
 # Release Notes
 
+## Version 1.1.7
+        * [CP-3399] - Fix incoming transactions not being displayed
+
+## Version 1.1.6
+        * [SER-3377] - Optimize Data Fetching Logic for PH Incoming Requests to Pay
+
+## Version 1.1.5
+        * [SER-3252] - Changed Currency selection field to dropdown on incoming request to pay filters
+
+## Version 1.1.4
+        * [SER-2248] - Resolve the issue on the export part and show user friendly error message when no records are found
+
+## Version 1.1.3
+        * [SER-3003] - Refactored Payer DFSP Id from free text to a dropdown and renamed to ‘AMS Business Short Code’
+
+## Version 1.1.2
+        * [SER-2070] - Show action logs
+
+## Version 1.1.1
+        * [SER-2991] - Add support for filtering NOT_AUTOSAVED incoming request to pay.
+
+## Version 1.1.0
+        * [SER-2909] - Add support for filtering payments by paybill short code and till number.
+        * [SER-2925] - Create UI for exporting incoming transactions.
+
+## Version 1.0.4
+        * [SER-2739] - Ensure transactions can be correctly filtered by currency and dates.
+
 ## Version 1.0.3
         * [SER-2208] - Fix bugs in searching incoming transactions in PaymentHub Service.
         * [SER-2257] - Standardize payment hub reports.